requirements.txt
Sphinx
sphinx-rtd-theme
git+https://github.com/ansible/ansible.git@devel
git+https://github.com/jameskyle/vmfusion-python.git@master
invoke
pytest
PyYAML
ansicolors
